feat: Update project URL and implement lazy loading for images
This commit updates the project URL for the React Quiz App in the projectsData.js file. The new URL is "https://quiz-app-772a7.web.app/". Additionally, it implements lazy loading for images in the Hero and Contact components, improving page load performance by delaying the loading of images until they are in the viewport.